+++ packages/vue/CHANGELOG.md
@@ -4,11 +4,10 @@ All notable changes to this project will be documented in this file.
 
 # [1.4.0](https://github.com/vime-js/vime/compare/@vime/vue@1.3.0...@vime/vue@1.4.0) (2020-08-25)
 
-
 ### Features
 
-* improve all framework integrations ([0ef1716](https://github.com/vime-js/vime/commit/0ef171655e8f02c277e0f00e90d87fdba8f74bb4))
-* new svelte store option ([ce78e15](https://github.com/vime-js/vime/commit/ce78e1532da638fffacdf6988ec66c9390c31f5d))
+- improve all framework integrations ([0ef1716](https://github.com/vime-js/vime/commit/0ef171655e8f02c277e0f00e90d87fdba8f74bb4))
+- new svelte store option ([ce78e15](https://github.com/vime-js/vime/commit/ce78e1532da638fffacdf6988ec66c9390c31f5d))
 
 # [1.3.0](https://github.com/vime-js/vime/compare/@vime/vue@1.2.0...@vime/vue@1.3.0) (2020-08-19)